py/django/imgscan/requirements.txt
appnope==0.1.2
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
black==22.1.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
executing==0.8.2
google-api-core==2.5.0
google-auth==2.6.0
google-cloud-vision==2.6.3
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
ipython==8.0.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
pluggy==1.0.0
prompt-toolkit==3.0.26
proto-plus==1.19.9
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
pyparsing==3.0.7
pytest==6.2.5
pytest-django==4.5.2
pytz==2021.3
requests==2.27.1
rsa==4.8
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
toml==0.10.2
tomli==2.0.0
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
